+++ src/test/java/org/folio/entlinks/service/consortium/ConsortiumAuthoritySourceFilePropagationServiceTest.java
@@ -1,8 +1,7 @@
 package org.folio.entlinks.service.consortium;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.folio.support.base.TestConstants.AUTHORITY_CONSORTIUM_SOURCE;
-import static org.folio.support.base.TestConstants.AUTHORITY_SOURCE;
+import static org.folio.entlinks.domain.entity.AuthoritySourceFileSource.CONSORTIUM;
 import static org.folio.support.base.TestConstants.TENANT_ID;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.doAnswer;
@@ -12,10 +11,11 @@ import static org.mockito.Mockito.when;
 
 import java.util.List;
 import java.util.UUID;
-import org.folio.entlinks.domain.entity.Authority;
+import org.folio.entlinks.domain.entity.AuthoritySourceFile;
 import org.folio.entlinks.exception.FolioIntegrationException;
-import org.folio.entlinks.service.authority.AuthorityService;
+import org.folio.entlinks.service.authority.AuthoritySourceFileService;
 import org.folio.entlinks.service.consortium.propagation.ConsortiumAuthorityPropagationService;
+import org.folio.entlinks.service.consortium.propagation.ConsortiumAuthoritySourceFilePropagationService;
 import org.folio.spring.service.SystemUserScopedExecutionService;
 import org.folio.spring.test.type.UnitTest;
 import org.junit.jupiter.api.Test;
@@ -27,68 +27,61 @@ import org.mockito.junit.jupiter.MockitoExtension;
 
 @UnitTest
 @ExtendWith(MockitoExtension.class)
-class ConsortiumAuthorityPropagationServiceTest {
+class ConsortiumAuthoritySourceFilePropagationServiceTest {
 
-  private @Mock AuthorityService authorityService;
+  private @Mock AuthoritySourceFileService authorityService;
   private @Mock ConsortiumTenantsService tenantsService;
   private @Mock SystemUserScopedExecutionService executionService;
-  private @InjectMocks ConsortiumAuthorityPropagationService propagationService;
+  private @InjectMocks ConsortiumAuthoritySourceFilePropagationService propagationService;
 
   @Test
   void testPropagateCreate() throws FolioIntegrationException {
-    var authority = authority();
-
+    var sourceFile = new AuthoritySourceFile();
     doMocks();
-    propagationService.propagate(authority, ConsortiumAuthorityPropagationService.PropagationType.CREATE, TENANT_ID);
+    propagationService.propagate(sourceFile, ConsortiumAuthorityPropagationService.PropagationType.CREATE, TENANT_ID);
 
-    assertThat(authority.getSource()).isEqualTo(AUTHORITY_CONSORTIUM_SOURCE);
+    assertThat(sourceFile.getSource()).isEqualTo(CONSORTIUM);
     verify(tenantsService).getConsortiumTenants(TENANT_ID);
     verify(executionService, times(3)).executeAsyncSystemUserScoped(any(), any());
-    verify(authorityService, times(3)).create(authority);
+    verify(authorityService, times(3)).create(sourceFile);
   }
 
   @Test
   void testPropagateUpdate() throws FolioIntegrationException {
-    var authority = authority();
-    authority.setId(UUID.randomUUID());
+    var sourceFile = new AuthoritySourceFile();
+    sourceFile.setId(UUID.randomUUID());
     doMocks();
-    propagationService.propagate(authority, ConsortiumAuthorityPropagationService.PropagationType.UPDATE, TENANT_ID);
+    propagationService.propagate(sourceFile, ConsortiumAuthorityPropagationService.PropagationType.UPDATE, TENANT_ID);
 
-    assertThat(authority.getSource()).isEqualTo(AUTHORITY_CONSORTIUM_SOURCE);
+    assertThat(sourceFile.getSource()).isEqualTo(CONSORTIUM);
     verify(tenantsService, times(1)).getConsortiumTenants(any());
     verify(executionService, times(3)).executeAsyncSystemUserScoped(any(), any());
-    verify(authorityService, times(3)).update(authority.getId(), authority);
+    verify(authorityService, times(3)).update(sourceFile.getId(), sourceFile);
   }
 
   @Test
   void testPropagateDelete() throws FolioIntegrationException {
-    var authority = authority();
-    authority.setId(UUID.randomUUID());
+    var sourceFile = new AuthoritySourceFile();
+    sourceFile.setId(UUID.randomUUID());
     doMocks();
-    propagationService.propagate(authority, ConsortiumAuthorityPropagationService.PropagationType.DELETE, TENANT_ID);
+    propagationService.propagate(sourceFile, ConsortiumAuthorityPropagationService.PropagationType.DELETE, TENANT_ID);
 
-    assertThat(authority.getSource()).isEqualTo(AUTHORITY_CONSORTIUM_SOURCE);
+    assertThat(sourceFile.getSource()).isEqualTo(CONSORTIUM);
     verify(tenantsService, times(1)).getConsortiumTenants(any());
     verify(executionService, times(3)).executeAsyncSystemUserScoped(any(), any());
-    verify(authorityService, times(3)).deleteById(authority.getId());
+    verify(authorityService, times(3)).deleteById(sourceFile.getId());
   }
 
   @Test
   void testPropagateException() throws FolioIntegrationException {
     Mockito.doThrow(FolioIntegrationException.class).when(tenantsService).getConsortiumTenants(any());
 
-    var authority = authority();
-    propagationService.propagate(authority, ConsortiumAuthorityPropagationService.PropagationType.CREATE, TENANT_ID);
+    var sourceFile = new AuthoritySourceFile();
+    propagationService.propagate(sourceFile, ConsortiumAuthorityPropagationService.PropagationType.CREATE, TENANT_ID);
 
     verify(tenantsService, times(1)).getConsortiumTenants(any());
     verify(executionService, times(0)).executeAsyncSystemUserScoped(any(), any());
     verify(authorityService, times(0)).create(any());
-  }
-
-  private Authority authority() {
-    var authority = new Authority();
-    authority.setSource(AUTHORITY_SOURCE);
-    return authority;
   }
 
   private void doMocks() {